+++ src/index.ts
@@ -6,5 +6,6 @@ export * from "./components/utilities/MapOrFallback.js";
 
 export * from "./components/Block.js";
 export * from "./components/Breadcrumbs.js";
+export * from "./components/ContentWrapper.js";
 export * from "./components/Header.js";
 export * from "./components/Notice.js";